% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{CSELR}
\alias{CSELR}
\alias{selr}
\title{Composite semiparametric empirical likelihood ratio}
\usage{
selr(dat)
}
\arguments{
\item{dat}{the data}
}
\value{
the value of Composite semiparametric empirical likelihood ratio
}
\description{
compute the composite semiparametric empirical likelihood value
}
\examples{
\dontrun{
s<-selr(data)
}
}
