% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(a, b, n)
}
\arguments{
\item{a}{the component elements of the first parameter of beta}

\item{b}{the component elements of the two parameter of beta}

\item{n}{the parameter of the binomial distribution}
}
\value{
a random sample
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC<-gibbsC(2,3,100)
plot(rnC[,1],type='l')
plot(rnC[,2],type='l')
}
}
