% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{GibbsR}
\alias{GibbsR}
\alias{gibbsR}
\title{A Gibbs sampler using R}
\usage{
gibbsR(a, b, n)
}
\arguments{
\item{a}{the component elements of the first parameter of beta}

\item{b}{the component elements of the two parameter of beta}

\item{n}{the parameter of the binomial distribution}
}
\value{
a random sample
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsR(100,2,3)
par(mfrow=c(2,1))
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
