\name{StatComp21027-package}
\alias{StatComp21027-package}
\alias{StatComp21027}
\docType{package}
\title{
  composite semiparametric empirical likelihood ratio in longitudinal data
}
\description{
  This R package provides users with two functions, mainly for testing the uniformity of the distribution of multiple sets of categorical data. The first function is to calculate the test statistics, and the second function provides the p value of the test through the bootstrap method.
}
\author{
Li Wenmei.

Maintainer: Li Wenmei <lwmqt@mail.ustc.edu.cn>
}
\keyword{ package }
