% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{statement}
\alias{statement}
\title{the statement of R packages we use}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{vaccR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
tm1<-microbenchmark::microbenchmark(
  rnR = gibbsR(2,3,100),
  rnC = gibbsC(2,3,100)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
